refactor: move the particle pooling and effects into the new particle field class
